Revert "convert comments"
This reverts commit c7f7e87c6cde2287f2e4549dc7d1ca7902538546.
